Remove extra::par fixes #5626
