+++ src/libextra/std.rc
@@ -100,7 +100,6 @@ pub mod tempfile;
 pub mod term;
 pub mod time;
 pub mod arena;
-pub mod par;
 pub mod base64;
 pub mod rl;
 pub mod workcache;
